+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ Changelog
 
 Release date: ``2020-xx-xx``
 
+- `NXPY-151 <https://jira.nuxeo.com/browse/NXPY-151>`__: Do not log the full exception when retrieving MPU parts
 - `NXPY-153 <https://jira.nuxeo.com/browse/NXPY-153>`__: Do not yield one more time only for S3 uploads
 
 Technical changes