+++ CHANGELOG.rst
@@ -9,6 +9,7 @@ Release date: ``2020-xx-xx``
 - `NXPY-151 <https://jira.nuxeo.com/browse/NXPY-151>`__: Do not log the full exception when retrieving MPU parts
 - `NXPY-152 <https://jira.nuxeo.com/browse/NXPY-152>`__: Remove Uploader assert statements
 - `NXPY-153 <https://jira.nuxeo.com/browse/NXPY-153>`__: Do not yield one more time only for S3 uploads
+- `NXPY-154 <https://jira.nuxeo.com/browse/NXPY-154>`__: Fix S3 client instanciation not thread-safe
 
 Technical changes
 -----------------